added start of getRestaurants for user controller
