+++ src/Restaurant/RestaurantModel.ts
@@ -1,13 +1,15 @@
 export class RestaurantModel{
     name='';
     tags:string[] = [];
+    events: any[] = [];
 
 
 
     static fromObject(object:any):RestaurantModel{
         const r:RestaurantModel=new RestaurantModel();
         r.name = object.name;
-        
+        r.tags = object.tags;
+        r.events = object.events;
         return r;
     }
